feat: convert .js to .ts after invoking
