+++ packages/@vue/cli-service/generator/vuex/index.js
@@ -7,4 +7,9 @@ module.exports = (api, options) => {
     }
   })
   api.render('./template')
+
+  if (api.invoking && api.hasPlugin('typescript')) {
+    const convertFiles = require('@vue/cli-plugin-typescript/generator/convert')
+    convertFiles(api)
+  }
 }
